+++ components/common/tests/var_registry/var_registry_link4.cpp
@@ -29,10 +29,10 @@ void register_handler (const char* mask, const VarRegistry& registry)
 {
   VarRegistryEvent events [] = {VarRegistryEvent_OnCreateVar, VarRegistryEvent_OnChangeVar, VarRegistryEvent_OnDeleteVar};
 
-  for (int i=0; i<sizeof (events)/sizeof (*events); i++)
+  for (size_t i = 0; i < sizeof (events) / sizeof (*events); i++)
   {
     VarRegistryEvent event = (VarRegistryEvent)i;
-    
+
     registry.RegisterEventHandler (mask, event, xtl::bind (&notify, _1, event, mask, xtl::cref (registry)));
   }
 }
@@ -40,21 +40,21 @@ void register_handler (const char* mask, const VarRegistry& registry)
 int main ()
 {
   printf ("Results of var_registry_link4_test:\n");
-  
+
   try
   {
     VarRegistrySystem::Mount ("a.b", TestVarRegistry::Create ().get ());
 
     VarRegistrySystem::Link ("a.c", "a.b");
-   
+
     VarRegistry registry ("a");
 
     register_handler ("*", registry);
-    
+
     registry.SetValue ("c.x", xtl::any (stl::string ("x"), true));
     registry.SetValue ("b.y", xtl::any (stl::string ("y"), true));
     registry.SetValue ("b.x.y", xtl::any (stl::string ("x.y"), true));
-    registry.SetValue ("c.xxxx.y", xtl::any (stl::string ("xxxx.y"), true));    
+    registry.SetValue ("c.xxxx.y", xtl::any (stl::string ("xxxx.y"), true));
 
     printf ("enumerate a vars:\n");
     registry.EnumerateVars (xtl::bind (&enumerator, _1, xtl::cref (registry)));
